pyproject.toml
[project]
name = "trading-app"
version = "0.1.0"
description = "Outil CLI de trading crypto futures"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ccxt>=4.5.37",
    "click>=8.3.1",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]

[project.scripts]
trading-app = "src.cli.main:cli"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
]
